Consultas addes

+++ src/main/java/com/aliengnss/backend/infraestructura/persistencia/jpa/ClienteJpa.java
@@ -23,7 +23,7 @@ public class ClienteJpa implements Serializable {
 	
 	@Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idCliente;
+    private Long idCliente;
